Use /home instead of /users for NFS share
For more commonplace openflight clusters it is likely that simple cloud
images will be used. Due to this, default/admin users will likely have
home directories under /home instead of /users thus making /home a
better default export to have and /users is something that can be
modified/added in by end-users when configuring the playbook.